Refactor deployments and add a dags path variable
This commit breaks the deployments out into separate files.
The purpose of this is for troubleshooting when making
changes to the deployment and fixing errors when finding
which deployment and line number is causing a problem.

This change also make the dag_path which is used for
volumeMounts within deployments.