Rename celery_result_backend to result_backend
After changes here [0] when running airflow with parameter
`celery_result_backend` scheduler and worker processes will crash
trying to connect to MySQL host, which is default host for celery.
This error is weird especially in case when celery is configured to
connect to PostgreSQL or different backend.
This patch will fix this issue.

[0] - https://github.com/apache/incubator-airflow/blob/master/UPDATING.md#celery-config